pass common constants defaultMenu , defaultMainColor, defaultEmails from ui/@/constants/general to startup/@/constants
