+++ imports/startup/server/installation.js
@@ -7,7 +7,7 @@ import {
   defaultMainColor, 
   defaultMenu, 
   defaultEmails 
-} from '../../ui/@/constants/general';
+} from '../../startup/@/constants';
 
 const username =  'ilkergonenc';
 const email =     'ilkergonenc@gmail.com';
